+++ api/src/Infrastructure/Persistence/Utilisateur/InMemoryUtilisateurReference.php
@@ -54,6 +54,14 @@ class InMemoryUtilisateurReference implements Utilisateur
   /**
    * {@inheritdoc}
    */
+  public function getMdp(): string
+  {
+    throw new ReferenceException();
+  }
+
+  /**
+   * {@inheritdoc}
+   */
   public function getNom(): string
   {
     throw new ReferenceException();