+++ api/tests/Application/Actions/Utilisateur/EditUtilisateurActionTest.php
@@ -6,6 +6,7 @@ namespace Tests\Application\Actions\Utilisateur;
 
 use App\Domain\Utilisateur\UtilisateurNotFoundException;
 use App\Infrastructure\Persistence\Utilisateur\DoctrineUtilisateur;
+use Prophecy\Argument;
 use Slim\Exception\HttpForbiddenException;
 use Slim\Exception\HttpNotFoundException;
 use Slim\Exception\HttpUnauthorizedException;
@@ -20,18 +21,27 @@ class EditUtilisateurActionTest extends ActionTestCase
             ->willReturn($this->alice)
             ->shouldBeCalledOnce();
 
-        /**
-         * @var DoctrineUtilisateur
-         */
+        $nouveauMdp = 'nouveaumdpalice';
+        /** @var DoctrineUtilisateur */
         $aliceModifiee = (new DoctrineUtilisateur($this->alice->getId()))
             ->setIdentifiant('alice2@tkdo.org')
             ->setNom('Alice2')
             ->setGenre('M')
-            ->setMdp('nouveaumdpalice')
+            ->setMdp(password_hash($nouveauMdp, PASSWORD_DEFAULT))
             ->setEstAdmin($this->alice->getEstAdmin());
+        $testCase = $this;
         $this->utilisateurRepositoryProphecy
-            ->update($aliceModifiee)
-            ->willReturn($aliceModifiee)
+            ->update(Argument::cetera())
+            ->will(function($args) use($testCase, $aliceModifiee, $nouveauMdp) {
+                /** @var Utilisateur */
+                $u = $args[0];
+                $testCase->assertEquals($aliceModifiee->getIdentifiant(), $u->getIdentifiant());
+                $testCase->assertEquals($aliceModifiee->getNom(), $u->getNom());
+                $testCase->assertEquals($aliceModifiee->getGenre(), $u->getGenre());
+                $testCase->assertEquals(true, password_verify($nouveauMdp, $u->getMdp()));
+                $testCase->assertEquals($aliceModifiee->getEstAdmin(), $u->getEstAdmin());
+                return $aliceModifiee;
+            })
             ->shouldBeCalledOnce();
 
         $estAdmin = json_encode($aliceModifiee->getEstAdmin());
@@ -45,7 +55,7 @@ class EditUtilisateurActionTest extends ActionTestCase
 {
     "genre": "{$aliceModifiee->getGenre()}",
     "identifiant": "{$aliceModifiee->getIdentifiant()}",
-    "mdp": "{$aliceModifiee->getMdp()}",
+    "mdp": "{$nouveauMdp}",
     "nom": "{$aliceModifiee->getNom()}",
     "estAdmin": $estAdmin
 }
@@ -53,7 +63,18 @@ class EditUtilisateurActionTest extends ActionTestCase
 EOT
         );
 
-        $this->assertEquals("null\n", (string)$response->getBody());
+        $estAdmin = json_encode($aliceModifiee->getEstAdmin());
+        $json = <<<EOT
+{
+    "estAdmin": $estAdmin,
+    "genre": "{$aliceModifiee->getGenre()}",
+    "id": {$aliceModifiee->getId()},
+    "identifiant": "{$aliceModifiee->getIdentifiant()}",
+    "nom": "{$aliceModifiee->getNom()}"
+}
+
+EOT;
+        $this->assertEquals($json, (string)$response->getBody());
     }
 
     public function testActionAdmin()
@@ -63,18 +84,28 @@ EOT
             ->willReturn($this->alice)
             ->shouldBeCalledOnce();
 
-        /**
-         * @var DoctrineUtilisateur
-         */
+        
+        $mdp = $this->mdpalice;
+        /** @var DoctrineUtilisateur */
         $aliceModifiee = (new DoctrineUtilisateur($this->alice->getId()))
             ->setIdentifiant('alice2@tkdo.org')
             ->setNom('Alice2')
             ->setGenre('M')
-            ->setMdp($this->alice->getMdp())
+            ->setMdp(password_hash($mdp, PASSWORD_DEFAULT))
             ->setEstAdmin(!$this->alice->getEstAdmin());
+        $testCase = $this;
         $this->utilisateurRepositoryProphecy
-            ->update($aliceModifiee)
-            ->willReturn($aliceModifiee)
+            ->update(Argument::cetera())
+            ->will(function ($args) use ($testCase, $aliceModifiee, $mdp) {
+                /** @var Utilisateur */
+                $u = $args[0];
+                $testCase->assertEquals($aliceModifiee->getIdentifiant(), $u->getIdentifiant());
+                $testCase->assertEquals($aliceModifiee->getNom(), $u->getNom());
+                $testCase->assertEquals($aliceModifiee->getGenre(), $u->getGenre());
+                $testCase->assertEquals(true, password_verify($mdp, $u->getMdp()));
+                $testCase->assertEquals($aliceModifiee->getEstAdmin(), $u->getEstAdmin());
+                return $aliceModifiee;
+            })
             ->shouldBeCalledOnce();
 
         $estAdmin = json_encode($aliceModifiee->getEstAdmin());
@@ -95,7 +126,18 @@ EOT
 EOT
         );
 
-        $this->assertEquals("null\n", (string)$response->getBody());
+        $estAdmin = json_encode($aliceModifiee->getEstAdmin());
+        $json = <<<EOT
+{
+    "estAdmin": $estAdmin,
+    "genre": "{$aliceModifiee->getGenre()}",
+    "id": {$aliceModifiee->getId()},
+    "identifiant": "{$aliceModifiee->getIdentifiant()}",
+    "nom": "{$aliceModifiee->getNom()}"
+}
+
+EOT;
+        $this->assertEquals($json, (string)$response->getBody());
     }
 
     public function testActionChangeEstAdmin()