+++ CHANGELOG.md
@@ -15,6 +15,7 @@
     - un accès à de nouvelles routes réservées aux administrateurs
       (et non accessibles par le front) :
       - GET et POST /api/utilisateur
+      - POST /api/utilisateur/:idUtilisateur/reinitmdp
   - côté front : l'accès à une page d'administration
     détaillant l'utilisation de ces routes en ligne de commande